Remove duplicate file
